Move internal miner/block creation to separate miner.cpp module.
Public functions referenced elsewhere are added to miner.h.
