+++ bitcoin-qt.pro
@@ -157,6 +157,7 @@ HEADERS += src/qt/bitcoingui.h \
     src/serialize.h \
     src/core.h \
     src/main.h \
+    src/miner.h \
     src/net.h \
     src/key.h \
     src/db.h \
@@ -240,6 +241,7 @@ SOURCES += src/qt/bitcoin.cpp \
     src/core.cpp \
     src/main.cpp \
     src/init.cpp \
+    src/miner.cpp \
     src/net.cpp \
     src/bloom.cpp \
     src/checkpoints.cpp \
